commit message

+++ 0x10-variadic_functions/0-sum_them_all.c
@@ -18,7 +18,7 @@ int sum_them_all(const unsigned int n, ...)
 
 		while (i < n)
 		{
-			sum += va_arg(args, int);
+			sum += va_arg(args, long long);
 			i++;
 		}
